Moved roster to the utils file
