+++ examtool/examtool/cli/deploy.py
@@ -5,10 +5,10 @@ import click
 from cryptography.fernet import Fernet
 
 from examtool.api.utils import rand_id
-from examtool.api.database import process_ok_exam_upload, set_exam, get_exam, set_roster, verify_roster
+from examtool.api.database import process_ok_exam_upload, set_exam, get_exam, set_roster
 from examtool.api.extract_questions import extract_questions, get_name
 from examtool.api.scramble import is_compressible_group, scramble
-from examtool.cli.utils import exam_name_option
+from examtool.cli.utils import exam_name_option, verify_roster
 
 
 @click.command()
