+++ examtool/examtool/cli/utils.py
@@ -22,6 +22,15 @@ hidden_target_folder_option = click.option(
 )
 
 
+def verify_roster(*, roster):
+    for i, row in enumerate(roster):
+        j = len(row)
+        if (j != 2):
+            print(f"ValueError: The roster must only contain 2 columns: Email, Timestamp. Found {j} item(s) on row {i + 1}: {row}")
+            return False
+    return True
+
+
 def prettify(course_code):
     m = re.match(r"([a-z]+)([0-9]+[a-z]?)", course_code)
     return m and (m.group(1) + " " + m.group(2)).upper()
